Move `colour.utilities.as_bool` definition to `colour.utilities.verbose` module.
